python-backend/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Agent 框架
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20

# 文档处理
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
Pillow==10.2.0
python-magic-bin==0.4.14  # Windows 文件类型检测

# Markdown 和 LaTeX 处理
markdown==3.5.1
python-markdown-math==0.8
mistune==3.0.2

# 工具库
pydantic==2.5.3
python-dotenv==1.0.0
